% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afit_presentation.R
\name{afit_presentation}
\alias{afit_presentation}
\title{Create an AFIT Presentation}
\usage{
afit_presentation(...)
}
\arguments{
\item{duration}{Duration (in minutes) of the slide deck. This value is used
to add a countdown timer to the slide footer.}

\item{footer}{Footer text (e.g. organization name and/or copyright)}

\item{font_adjustment}{Increase or decrease the default font size
(e.g. -1 or +1). You can also manually adjust the font size during the
presentation using the 'S' (smaller) and 'B' (bigger) keys.}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a slidy presentation.
}
\details{
See the \href{http://rmarkdown.rstudio.com/slidy_presentation_format.html}{online
documentation} for additional details on using the \code{slidy_presentation} format.

For more information on markdown syntax for presentations see the
\href{http://pandoc.org/README.html}{pandoc online documentation}.
}
\examples{
\dontrun{

library(AFIT)

# simple invocation
render("afit_presentation.Rmd", slidy_presentation())

# specify an option for incremental rendering
render("afit_presentation.Rmd", slidy_presentation(incremental = TRUE))
}

}

