% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afit_thesis.R
\name{afit_thesis}
\alias{afit_thesis}
\title{Air Force Institute of Technology (AFIT) Thesis format.}
\usage{
afit_thesis(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating AFIT Theses.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "afit_thesis", package = "AFIT")
}

}

