% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{forms}
\alias{forms}
\title{Open AFIT-Specific forms in a browser}
\usage{
forms(form.name = NULL)
}
\arguments{
\item{form.name}{The title of the requested form}
}
\description{
This function from to complete an AFIT Thesis, the
}

