% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noTouch.R
\name{noTouch}
\alias{noTouch}
\title{Extract YAML data from metadata.yml}
\usage{
noTouch(file = NULL, item = NULL, ...)
}
\arguments{
\item{item}{The YAML item to be extracted}

\item{up}{If \code{TRUE} the search path for the file moves up one level}
}
\description{
Extract YAML data from metadata.yml
}

