% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preamble.R
\name{preamble}
\alias{preamble}
\title{Air Force Institute of Technology Preamble}
\usage{
preamble(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"),
  fig_caption = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{fig_caption}{\code{TRUE} to render figures with captions}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating the initial pages of AFIT theses, dissertations, and reports.
Adapted from the STYLE GUIDE FOR AFIT DISSERTATIONS, THESES AND GRADUATE RESEARCH PAPERS - July 2015
\href{http://www.afit.edu/docs/NewAFITStyleGuideJuly2015(2).pdf}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyThesis.Rmd", template = "preamble", package = "AFIT")
}

}

